requirements.txt
Django==5.1.7
djangorestframework==3.15.2
requests==2.32.3
psycopg2-binary==2.9.10
psycopg2==2.9.10